Fixed up app.jsx
