+++ server/src/App.jsx
@@ -1,23 +1,38 @@
-import { useState } from 'react'
 import './App.css'
-import GameCard from './Components/GameCard';
-import SearchBar from './Components/SearchBar';
+import { useState } from 'react'
+import axios from 'axios'
 
 function App() {
-  const [count, setCount] = useState(0)
+const [gameTitle, setGameTitle] = useState('');
+const [results, setResults] = useState([]);
+
+const handleSearch = async () => {
+  try {
+    const response = await axios.post('https://video-game-search-forum.onrender.com/api/search', {query: gameTitle});
+    setResults(response.data);
+  } catch (error) {
+    console.log('Oh god, not an error!:', error);
+  }
+};
 
   return (
-   
-      <div className="page">
-        <header>
-          <h1>GAMEFO</h1>
-          <SearchBar/>
-        </header>
-        <GameCard />
-        
-      </div>
-    
-  )
+    <div>
+      <h1>Search for a game</h1>
+      <input
+        type='text'
+        value={gameTitle}
+        onChange={(e) => setGameTitle(e.target.value)}
+        placeholder='Search for a game'
+        />
+        <button onClick={handleSearch}>Search</button>
+
+        <ul>
+          {results.map((game) => (
+            <li key={game.id}>{game.name}</li>
+          ))}
+        </ul>
+    </div>
+  );
 }
 
 export default App
