--- src/routes/userRoutes.ts
@@ -1,26 +1,0 @@
-/* import express from 'express':
-import User from '../models/User';
-
-const router = express.Router();
-
-//Create a new User
-router.post('/users', async (req, res) => {
-    try {
-        const user = await User.create(req.body);
-        res.status(201).json(user);
-    } catch (err) {
-        res.status(500).json({ error: err.message });
-    }
-    });
-
-// Fetch all users
-router.get('/users', async (req, res) => {
-    try {
-        const users = await User.findAll();
-        res.status(200).json(users);
-    } catch (err) {
-        res.status(500).json({ error: err.message});
-    }
-});
-
-export default router; */
